+++ CHANGES.rst
@@ -7,6 +7,8 @@ tsfresh uses `Semantic Versioning <http://semver.org/>`_
 Unreleased
 ==========
 
+We changed the default branch from "master" to "main".
+
 - Breaking Change
     - Changed constructed id in roll_time_series from string to tuple (#700)
     - Same for add_sub_time_series_index (#720)
@@ -20,6 +22,7 @@ Unreleased
     - Speed up the result pivoting (#705)
     - Add a test for the dask bindings (#719)
     - Refactor input data iteration to need less memory (#707)
+    - Added benchmark tests (#710)
 - Bugfixes:
     - Fixed a bug in the selection, that caused all regression tasks with un-ordered index to be wrong (#715)
     - Fixed readthedocs (#695, #696)